feat: added scope and collection API support

+++ android/src/main/java/com/cblreactnative/CblReactnativeModule.kt
@@ -1,6 +1,7 @@
 package com.cblreactnative
 
 import cbl.js.kotiln.DatabaseManager
+import cbl.js.kotiln.CollectionManager
 import cbl.js.kotiln.FileSystemHelper
 import cbl.js.kotiln.LoggingManager
 import com.couchbase.lite.*
@@ -36,6 +37,7 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
   }
 
   // Collection Functions
+
   @ReactMethod
   fun collection_CreateCollection(
     collectionName: String,
@@ -43,6 +45,9 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
     scopeName: String,
     promise: Promise) {
     try {
+      if(!DataValidation.validateCollection(collectionName, scopeName, name, promise)) {
+        return
+      }
       val col = DatabaseManager.createCollection(collectionName, scopeName, name)
       col?.let { collection ->
         val colMap = DataAdapter.adaptCollectionToMap(collection, name)
@@ -62,6 +67,9 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
     scopeName: String,
     promise: Promise) {
     try {
+      if(!DataValidation.validateCollection(collectionName, scopeName, name, promise)) {
+        return
+      }
       DatabaseManager.deleteCollection(collectionName, scopeName, name)
       promise.resolve(null)
     } catch (e: Exception) {
@@ -76,6 +84,9 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
     scopeName: String,
     promise: Promise) {
     try {
+      if(!DataValidation.validateCollection(collectionName, scopeName, name, promise)) {
+        return
+      }
       val col = DatabaseManager.getCollection(collectionName, scopeName, name)
       col?.let { collection ->
         val colMap = DataAdapter.adaptCollectionToMap(collection, name)
@@ -88,14 +99,75 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
     }
   }
 
+  @ReactMethod
+  fun collection_GetCollections(
+    name: String,
+    scopeName: String,
+    promise: Promise) {
+    try {
+      if (!DataValidation.validateScope(scopeName, name, promise)) {
+        return
+      }
+      val cols = DatabaseManager.getCollections(scopeName, name)
+      val colList = Arguments.createArray()
+      cols?.forEach { collection ->
+        val colMap = DataAdapter.adaptCollectionToMap(collection, name)
+        colList.pushMap(colMap)
+      }
+      val resultsCollections:WritableMap = Arguments.createMap()
+      resultsCollections.putArray("collections", colList)
+      promise.resolve(resultsCollections)
+    } catch (e: Exception) {
+      promise.reject("DATABASE_ERROR", e.message)
+    }
+  }
+
+  @ReactMethod
+  fun collection_GetCount(
+    collectionName: String,
+    name: String,
+    scopeName: String,
+    promise: Promise) {
+    try {
+      if(!DataValidation.validateCollection(collectionName, scopeName, name, promise)) {
+        return
+      }
+      val count = CollectionManager.documentsCount(collectionName, scopeName, name)
+      val map = Arguments.createMap()
+      map.putInt("count", count)
+      promise.resolve(map)
+    } catch (e: Exception) {
+      promise.reject("DATABASE_ERROR", e.message)
+    }
+  }
+
+  @ReactMethod
+  fun collection_GetDefault(
+    name: String,
+    promise: Promise){
+    if (!DataValidation.validateDatabaseName(name, promise)) {
+      return
+    }
+    try {
+      val col = DatabaseManager.defaultCollection(name)
+      col?.let { collection ->
+        val colMap = DataAdapter.adaptCollectionToMap(collection, name)
+        promise.resolve(colMap)
+        return
+      }
+      promise.reject("COLLECTION_ERROR", "Error getting default collection")
+    } catch (e: Exception) {
+      promise.reject("DATABASE_ERROR", e.message)
+    }
+  }
+
   // Database Functions
   @ReactMethod
   fun database_ChangeEncryptionKey(
     newKey: String,
     name: String,
     promise: Promise) {
-    if (name.isEmpty()) {
-      promise.reject("DATABASE_ERROR", "Database name must be provided")
+    if (!DataValidation.validateDatabaseName(name, promise)) {
       return
     }
     try {
@@ -110,8 +182,7 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
   fun database_Close(
     name: String,
     promise: Promise) {
-    if (name.isEmpty()) {
-      promise.reject("DATABASE_ERROR", "Database name must be provided")
+    if (!DataValidation.validateDatabaseName(name, promise)) {
       return
     }
     try {
@@ -131,12 +202,10 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
     encryptionKey: String?,
     promise: Promise) {
     try {
-      if (name.isEmpty()) {
-        promise.reject("DATABASE_ERROR", "Database name must be provided")
+      if (!DataValidation.validateDatabaseName(newName, promise)) {
         return
       }
-      if (path.isEmpty()) {
-        promise.reject("DATABASE_ERROR", "Database path must be provided")
+      if (!DataValidation.validatePath(path, promise)) {
         return
       }
       val databaseConfig = DataAdapter.getDatabaseConfig(directory, encryptionKey)
@@ -151,8 +220,7 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
   fun database_Delete(
     name: String,
     promise: Promise) {
-    if (name.isEmpty()) {
-      promise.reject("DATABASE_ERROR", "Database name must be provided")
+    if (!DataValidation.validateDatabaseName(name, promise)) {
       return
     }
     try {
@@ -168,12 +236,10 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
     path: String,
     name: String,
     promise: Promise) {
-    if (name.isEmpty()) {
-      promise.reject("DATABASE_ERROR", "Database name must be provided")
+    if (!DataValidation.validateDatabaseName(name, promise)) {
       return
     }
-    if (path.isEmpty()) {
-      promise.reject("DATABASE_ERROR", "Database path must be provided")
+    if (!DataValidation.validatePath(path, promise)) {
       return
     }
     try {
@@ -189,8 +255,10 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
     name: String,
     directory: String,
     promise: Promise) {
-    if (name.isEmpty()) {
-      promise.reject("DATABASE_ERROR", "Database name must be provided")
+    if (!DataValidation.validateDatabaseName(name, promise)) {
+      return
+    }
+    if (!DataValidation.validatePath(directory, promise)) {
       return
     }
     try {
@@ -205,8 +273,7 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
   fun database_GetPath(
     name: String,
     promise: Promise) {
-    if (name.isEmpty()) {
-      promise.reject("DATABASE_ERROR", "Database name must be provided")
+    if (!DataValidation.validateDatabaseName(name, promise)) {
       return
     }
     try {
@@ -223,8 +290,7 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
     directory: String? = null,
     encryptionKey: String? = null,
     promise: Promise) {
-    if (name.isEmpty()) {
-      promise.reject("DATABASE_ERROR", "Database name must be provided")
+    if (!DataValidation.validateDatabaseName(name, promise)) {
       return
     }
     try {
@@ -244,8 +310,7 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
     maintenanceType: Double,
     databaseName: String,
     promise: Promise) {
-    if (databaseName.isEmpty()) {
-      promise.reject("DATABASE_ERROR", "Database name must be provided")
+    if (!DataValidation.validateDatabaseName(name, promise)) {
       return
     }
     try {
@@ -282,6 +347,74 @@ class CblReactnativeModule(reactContext: ReactApplicationContext) :
       promise.resolve(null)
     } catch (e: Exception){
       promise.reject("LOGGING_ERROR", e.message)
+    }
+  }
+
+  // Scope Functions
+  @ReactMethod
+  fun scope_GetDefault(
+    name: String,
+    promise: Promise ){
+   try {
+      if (!DataValidation.validateDatabaseName(name, promise)) {
+        return
+      }
+      val scopeValue = DatabaseManager.defaultScope(name)
+      scopeValue?.let { scope ->
+        val scopeMap = DataAdapter.adaptScopeToMap(scope, name)
+        promise.resolve(scopeMap)
+        return
+      }
+      promise.reject("SCOPE_ERROR", "Error getting default scope")
+   } catch (e: Exception) {
+     promise.reject("SCOPE_ERROR", e.message)
+   }
+  }
+
+  @ReactMethod
+  fun scope_GetScope(
+    scopeName: String,
+    name: String,
+    promise: Promise) {
+    try {
+      if (!DataValidation.validateScope(scopeName, name, promise)) {
+        return
+      }
+      val scopeValue = DatabaseManager.getScope(name, scopeName)
+      scopeValue?.let { scope ->
+        val scopeMap = DataAdapter.adaptScopeToMap(scope, name)
+        promise.resolve(scopeMap)
+        return
+      }
+      promise.reject("SCOPE_ERROR", "Error getting scope")
+    } catch (e: Exception) {
+      promise.reject("SCOPE_ERROR", e.message)
+    } catch(e: Exception) {
+      promise.reject("SCOPE_ERROR", e.message)
+    }
+  }
+
+  @ReactMethod
+  fun scope_GetScopes(
+    name: String,
+    promise: Promise){
+    try {
+      if (!DataValidation.validateDatabaseName(name, promise)) {
+        return
+      }
+      val scopes = DatabaseManager.scopes(name)
+      val scopeList = Arguments.createArray()
+      scopes.forEach { scope ->
+        val scopeMap = DataAdapter.adaptScopeToMap(scope, name)
+        scopeList.pushMap(scopeMap)
+      }
+      val resultsScopes:WritableMap = Arguments.createMap()
+      resultsScopes.putArray("scopes", scopeList)
+      promise.resolve(resultsScopes)
+    } catch (e: Exception) {
+      promise.reject("SCOPE_ERROR", e.message)
+    } catch (e: Exception) {
+      promise.reject("SCOPE_ERROR", e.message)
     }
   }
